fix: relative code calling
If you change the `call[rsp-24]` to `call[rsp-32]` then the simple tests
should all work.  It bypasses all the extra stuff that multiverse has as
scaffolding, so other things that actually need it might not work.

It seems there's a lot of stuff that actually needs to get the relative
calling fix.  So we probably need a more robust solution for this.