fix: account for symbols
Symbols exist after the last LOAD segment, so we need to make sure all
of our new segments are placed after that.

The Program Header segment has a mix of physical and virtual addressing,
hopefully nothing goes wrong there.

The popgm code I'm not sure how it worked to begin with, possibly it
just happened to fall into a proper memory address or maybe the gm table
used to be after the popgm code.

There's also a reduction in the unsupported instructions for RIP
adjustment in here.